requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pytest
httpx
python-dotenv
